refactor: apollo middleware
