+++ server/src/middleware/apollo.js
@@ -1,0 +1,92 @@
+// @ts-check
+const { expressMiddleware } = require('@apollo/server/express4')
+const { GraphQLError } = require('graphql')
+const { ApolloServerErrorCode } = require('@apollo/server/errors')
+const { parse } = require('graphql')
+
+const { Db, Event } = require('../services/initialization')
+const pkg = require('../../../package.json')
+
+/**
+ *
+ * @param {Awaited<ReturnType<import('../graphql/server')>>} server
+ * @returns
+ */
+function apolloMiddleware(server) {
+  return expressMiddleware(server, {
+    context: async ({ req, res }) => {
+      const perms = req.user ? req.user.perms : req.session.perms
+      const user = req?.user?.username || ''
+      const id = req?.user?.id || 0
+
+      const clientVHeader = req.headers['apollographql-client-version']
+      const clientV =
+        (typeof clientVHeader === 'string' && clientVHeader.trim()) ||
+        pkg.version ||
+        1
+      const serverV = pkg.version || 1
+
+      const definition =
+        /** @type {import('graphql').OperationDefinitionNode} */ (
+          parse(req.body.query).definitions.find(
+            (d) => d.kind === 'OperationDefinition',
+          )
+        )
+      const endpoint = definition?.name?.value || ''
+      const errorCtx = {
+        id,
+        user,
+        clientV,
+        serverV,
+        endpoint,
+      }
+
+      if (clientV && serverV && clientV !== serverV) {
+        throw new GraphQLError('old_client', {
+          extensions: {
+            ...errorCtx,
+            http: { status: 464 },
+            code: ApolloServerErrorCode.BAD_USER_INPUT,
+          },
+        })
+      }
+
+      if (!perms && endpoint !== 'Locales') {
+        throw new GraphQLError('session_expired', {
+          extensions: {
+            ...errorCtx,
+            http: { status: 511 },
+            code: 'EXPIRED',
+          },
+        })
+      }
+
+      if (
+        definition?.operation === 'mutation' &&
+        !id &&
+        endpoint !== 'SetTutorial'
+      ) {
+        throw new GraphQLError('unauthenticated', {
+          extensions: {
+            ...errorCtx,
+            http: { status: 401 },
+            code: 'UNAUTHENTICATED',
+          },
+        })
+      }
+
+      return {
+        req,
+        res,
+        Db,
+        Event,
+        perms,
+        user,
+        token: req.headers.token,
+        operation: definition?.operation,
+      }
+    },
+  })
+}
+
+module.exports = { apolloMiddleware }
